// hr-frontend/src/lib/api/generated/model/unitPosition.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export interface UnitPosition {
  id?: number;
  title?: string;
  code?: string;
  grade?: string;
  department?: string;
  reportsTo?: string;
  status?: string;
  employee?: string;
}